requirements.txt
gtts == 2.5.4
python-telegram-bot == 21.6
python-decouple == 3.8
python-docx == 1.1.2
PyPDF2 == 3.0.1